backend/pyproject.toml
[tool.poetry]
name = "iot-support-backend"
version = "0.1.0"
description = "Flask backend for managing ESP32 IoT device configurations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.scripts]
dev = "run:main"
iotsupport-cli = "app.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
spectree = "^1.2.0"
flask-cors = "^4.0.0"
waitress = "^3.0.0"
dependency-injector = "^4.48.1"
prometheus-flask-exporter = "^0.23.0"
cryptography = "^42.0.0"
pillow = "^11.0.0"
httpx = "^0.27.0"
pypng = "^0.20220715.0"
lz4 = "^4.3.0"
paho-mqtt = "^2.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
psycopg = {extras = ["binary"], version = "^3.1.0"}
alembic = "^1.13.0"
croniter = "^6.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
mypy = "^1.8.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
types-flask = "^1.1.0"
types-flask-cors = "^4.0.0"
types-waitress = "^3.0.1.20250801"
types-croniter = "^6.0.0.20250809"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"app/utils/lvgl/LVGLImage.py" = ["I001", "UP035", "B904", "UP006", "C416", "B905", "E741"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_errors = true

[[tool.mypy.overrides]]
module = "spectree.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "prometheus_client.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "png.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lz4.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "PIL.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "paho.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "flask_sqlalchemy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sqlalchemy", "sqlalchemy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.utils.lvgl.LVGLImage"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::DeprecationWarning:pydantic.*"
]

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]
